+++ apps/studio/src/common/interfaces/IMenuActionHandler.ts
@@ -1,6 +1,7 @@
 // import { OpenOptions } from "@/background/WindowBuilder"
 
 import { DevLicenseState } from "@/lib/license";
+import { CustomMenuAction } from "@/types";
 
 type ElectronWindow = Electron.BrowserWindow | undefined
 
@@ -45,4 +46,7 @@ export interface IMenuActionHandler {
   toggleBeta: (menuItem: Electron.MenuItem, win: ElectronWindow) => void
   managePlugins: (menuItem: Electron.MenuItem, win: ElectronWindow) => void
   updatePin: (menuItem: Electron.MenuItem, win: ElectronWindow) => void
+  /** `handleAction` is used by menus that are defined at runtime (unlike other
+   * actions) so the signature is a little bit different than the rest. */
+  handleAction: (action: Electron.MenuItem | CustomMenuAction, win: ElectronWindow) => void
 }
