+++ apps/studio/src/lib/menu/ClientMenuActionHandler.ts
@@ -2,6 +2,7 @@ import { IMenuActionHandler } from '@/common/interfaces/IMenuActionHandler'
 import _ from 'lodash'
 import {AppEvent} from '../../common/AppEvent'
 import rawLog from '@bksLogger'
+import { CustomMenuAction } from '@/types'
 
 const log = rawLog.scope("ClientMenuActionHandler")
 
@@ -61,6 +62,5 @@ export default class ClientMenuActionHandler implements IMenuActionHandler {
   }
   updatePin = () => send('updatePin')
   managePlugins = () => send("managePlugins")
-
-  send = send
+  handleAction = (action: CustomMenuAction) => send('handleAction', action)
 }
