+++ apps/studio/src/types.ts
@@ -1,3 +1,5 @@
+import { AppEvent } from "./common/AppEvent";
+
 interface UtilProcReadyMessage {
   type: "ready";
 }
@@ -12,3 +14,17 @@ export type UtilProcMessage =
   | UtilProcOpenExternalMessage;
 
 export type JsonValue = string | number | boolean | null | JsonValue[] | { [key: string]: JsonValue };
+
+export type CustomMenuAction = {
+  event: AppEvent;
+  args?: JsonValue;
+}
+
+export type ExternalMenuItem = {
+  id: string;
+  parentId: string;
+  label: string;
+  enableWhenConnected?: boolean;
+  action: CustomMenuAction;
+};
+
